requirements.txt
openai>=1.0.0
pyyaml
evdev
pyperclip
pyinstaller